Restructure blueprints, add flask-login
