+++ pyproject.toml
@@ -13,6 +13,7 @@ Flask-SQLAlchemy = "^2.5.1"
 mysql-connector-python = "^8.0.28"
 Flask-Admin = "^1.6.0"
 sqlacodegen = "^2.3.0"
+Flask-Login = "^0.6.0"
 
 [tool.poetry.dev-dependencies]
